# Batch NewLeaf node work in base API