--- a/conversion.go
+++ b/conversion.go
@@ -4,24 +4,21 @@ import (
 	"bytes"
 	"fmt"
 	"runtime"
-	"sort"
 	"sync"
 )
 
-// BatchNewLeafNodeData is a struct that contains the data needed to create a new leaf node.
-type BatchNewLeafNodeData struct {
-	Stem   []byte
-	Values map[byte][]byte
-}
-
-// BatchNewLeafNode creates a new leaf node from the given data. It optimizes LeafNode creation
-// by batching expensive cryptography operations. It returns the LeafNodes sorted by stem.
-func BatchNewLeafNode(nodesValues []BatchNewLeafNodeData) []LeafNode {
-	cfg := GetConfig()
-	ret := make([]LeafNode, len(nodesValues))
+func batchCommitLeafNodes(leaves []*LeafNode) {
+	minBatchSize := 4
+	if len(leaves) < minBatchSize {
+		commitLeafNodes(leaves)
+		return
+	}
 
 	numBatches := runtime.NumCPU()
-	batchSize := len(nodesValues) / numBatches
+	if numBatches > (len(leaves)+minBatchSize-1)/minBatchSize {
+		numBatches = (len(leaves) + minBatchSize - 1) / minBatchSize
+	}
+	batchSize := len(leaves) / numBatches
 
 	var wg sync.WaitGroup
 	wg.Add(numBatches)
@@ -29,46 +26,60 @@ func BatchNewLeafNode(nodesValues []BatchNewLeafNodeData) []LeafNode {
 		start := i * batchSize
 		end := (i + 1) * batchSize
 		if i == numBatches-1 {
-			end = len(nodesValues)
+			end = len(leaves)
 		}
-		go func(ret []LeafNode, nodesValues []BatchNewLeafNodeData) {
+		go func(leaves []*LeafNode) {
 			defer wg.Done()
-
-			c1c2points := make([]*Point, 2*len(nodesValues))
-			c1c2frs := make([]*Fr, 2*len(nodesValues))
-			for i, nv := range nodesValues {
-				valsslice := make([][]byte, NodeWidth)
-				for idx := range nv.Values {
-					valsslice[idx] = nv.Values[idx]
-				}
-
-				ret[i] = *NewLeafNode(nv.Stem, valsslice)
-
-				c1c2points[2*i], c1c2points[2*i+1] = ret[i].c1, ret[i].c2
-				c1c2frs[2*i], c1c2frs[2*i+1] = new(Fr), new(Fr)
-			}
-
-			toFrMultiple(c1c2frs, c1c2points)
-
-			var poly [NodeWidth]Fr
-			poly[0].SetUint64(1)
-			for i, nv := range nodesValues {
-				StemFromBytes(&poly[1], nv.Stem)
-				poly[2] = *c1c2frs[2*i]
-				poly[3] = *c1c2frs[2*i+1]
-
-				ret[i].commitment = cfg.CommitToPoly(poly[:], 252)
-			}
-
-		}(ret[start:end], nodesValues[start:end])
+			commitLeafNodes(leaves)
+		}(leaves[start:end])
 	}
 	wg.Wait()
+}
 
-	sort.Slice(ret, func(i, j int) bool {
-		return bytes.Compare(ret[i].stem, ret[j].stem) < 0
-	})
+func commitLeafNodes(leaves []*LeafNode) {
+	cfg := GetConfig()
 
-	return ret
+	c1c2points := make([]*Point, 2*len(leaves))
+	c1c2frs := make([]*Fr, 2*len(leaves))
+	for i, n := range leaves {
+		// C1.
+		var c1poly [NodeWidth]Fr
+		count := fillSuffixTreePoly(c1poly[:], n.values[:NodeWidth/2])
+		containsEmptyCodeHash := len(c1poly) >= EmptyCodeHashSecondHalfIdx &&
+			c1poly[EmptyCodeHashFirstHalfIdx].Equal(&EmptyCodeHashFirstHalfValue) &&
+			c1poly[EmptyCodeHashSecondHalfIdx].Equal(&EmptyCodeHashSecondHalfValue)
+		if containsEmptyCodeHash {
+			// Clear out values of the cached point.
+			c1poly[EmptyCodeHashFirstHalfIdx] = FrZero
+			c1poly[EmptyCodeHashSecondHalfIdx] = FrZero
+			// Calculate the remaining part of c1 and add to the base value.
+			partialc1 := cfg.CommitToPoly(c1poly[:], NodeWidth-count-2)
+			n.c1 = new(Point)
+			n.c1.Add(&EmptyCodeHashPoint, partialc1)
+		} else {
+			n.c1 = cfg.CommitToPoly(c1poly[:], NodeWidth-count)
+		}
+
+		// C2.
+		var c2poly [NodeWidth]Fr
+		count = fillSuffixTreePoly(c2poly[:], n.values[NodeWidth/2:])
+		n.c2 = cfg.CommitToPoly(c2poly[:], NodeWidth-count)
+
+		c1c2points[2*i], c1c2points[2*i+1] = n.c1, n.c2
+		c1c2frs[2*i], c1c2frs[2*i+1] = new(Fr), new(Fr)
+	}
+
+	toFrMultiple(c1c2frs, c1c2points)
+
+	var poly [NodeWidth]Fr
+	poly[0].SetUint64(1)
+	for i, nv := range leaves {
+		StemFromBytes(&poly[1], nv.stem)
+		poly[2] = *c1c2frs[2*i]
+		poly[3] = *c1c2frs[2*i+1]
+
+		nv.commitment = cfg.CommitToPoly(poly[:], 252)
+	}
 }
 
 // firstDiffByteIdx will return the first index in which the two stems differ.