This PR removes the new-ish specialized API for inserting new leaves that we created in https://github.com/gballet/go-verkle/pull/343, and does some internal refactors to have equal performance just using the base API (i.e: `Insert(key, value)`).

The high-level idea is to avoid doing CPU-heavy work in `NewLeafNode(...)` and let the `commitment` be nil. We add awareness in some code sections that case can happen, and simply means that’s an uncommited new leaf node.

In particular, when we call:

- `(*InternalNode).Commit(...)`: we stop assuming the leafs *are prepared*. Instead, we detect all *uncommited* leaves and do the batch-way of creating a set of uncommited leaves that I created in https://github.com/gballet/go-verkle/pull/343 but with some twists (but the same idea).
- `(LeafNode).Commit(...)`: we only commit that `LeafNode` using the same batch method but with a single element; so the logic is reduced.

I also refined the way the new-leaf-node-batching works but creating hot-paths for single-leaf node creation and only going full steam using cores if there’s enough work to do. (More about this in PR comments).

Note that we aren’t doing exactly https://github.com/gballet/go-verkle/pull/314. There’s no COW here in leaves since we’re exclusively focusing on optimizing freshly created leaves which is exactly the case for overlay tree key/value migrations. But, this optimization will also be exploited under normal block-execution circumstances since new/fresh leaves will be created too.

The logic for updating values in existing leaves remains the same as today; doing diff updating. If we ever want to optimize this case, *then* it might make sense to introduce the COW idea. But that needs more justification since it introduces a new map and more logical complexity. This PR doesn’t introduce extra memory overhead.

Below are the comparisons between the current-newish batch API for key/value migration and the new version of `Insert(key, value)`.

## Synthetic insertion

As a refresher, this was a simulation of a pre-existing tree with X random key/values, and then how long it takes to insert another random Y key/values.

Below I repeat the output of our current `master` branch. Note that the important number is `batched XXms` since that is showing how fast the new APIs that we introduced runs:

```bash
Assuming 0 key/values touched by block execution:
        If 1000 extra key-values are migrated: unbatched 43ms, batched 23ms, 1.89x
        If 2000 extra key-values are migrated: unbatched 78ms, batched 39ms, 2.01x
        If 5000 extra key-values are migrated: unbatched 192ms, batched 87ms, 2.21x
        If 8000 extra key-values are migrated: unbatched 304ms, batched 137ms, 2.22x
Assuming 500 key/values touched by block execution:
        If 1000 extra key-values are migrated: unbatched 44ms, batched 29ms, 1.50x
        If 2000 extra key-values are migrated: unbatched 81ms, batched 44ms, 1.82x
        If 5000 extra key-values are migrated: unbatched 193ms, batched 95ms, 2.03x
        If 8000 extra key-values are migrated: unbatched 308ms, batched 144ms, 2.14x
Assuming 1000 key/values touched by block execution:
        If 1000 extra key-values are migrated: unbatched 51ms, batched 34ms, 1.47x
        If 2000 extra key-values are migrated: unbatched 88ms, batched 52ms, 1.67x
        If 5000 extra key-values are migrated: unbatched 200ms, batched 102ms, 1.95x
        If 8000 extra key-values are migrated: unbatched 317ms, batched 150ms, 2.10x
Assuming 2000 key/values touched by block execution:
        If 1000 extra key-values are migrated: unbatched 63ms, batched 47ms, 1.32x
        If 2000 extra key-values are migrated: unbatched 99ms, batched 63ms, 1.58x
        If 5000 extra key-values are migrated: unbatched 214ms, batched 113ms, 1.88x
        If 8000 extra key-values are migrated: unbatched 328ms, batched 165ms, 1.99x
Assuming 5000 key/values touched by block execution:
        If 1000 extra key-values are migrated: unbatched 101ms, batched 86ms, 1.17x
        If 2000 extra key-values are migrated: unbatched 137ms, batched 99ms, 1.38x
        If 5000 extra key-values are migrated: unbatched 252ms, batched 151ms, 1.67x
        If 8000 extra key-values are migrated: unbatched 370ms, batched 204ms, 1.81x
```

Below, is the same test/benchmark but in this PR. i.e: using `Insert(key, value)` , what we called “unbatched” above. (i.e: no special APIs used):

```bash
Assuming 0 key/values touched by block execution:
        If 1000 extra key-values are migrated: unbatched 23ms
        If 2000 extra key-values are migrated: unbatched 38ms
        If 5000 extra key-values are migrated: unbatched 85ms
        If 8000 extra key-values are migrated: unbatched 132ms
Assuming 500 key/values touched by block execution:
        If 1000 extra key-values are migrated: unbatched 29ms
        If 2000 extra key-values are migrated: unbatched 44ms
        If 5000 extra key-values are migrated: unbatched 92ms
        If 8000 extra key-values are migrated: unbatched 139ms
Assuming 1000 key/values touched by block execution:
        If 1000 extra key-values are migrated: unbatched 36ms
        If 2000 extra key-values are migrated: unbatched 52ms
        If 5000 extra key-values are migrated: unbatched 100ms
        If 8000 extra key-values are migrated: unbatched 146ms
Assuming 2000 key/values touched by block execution:
        If 1000 extra key-values are migrated: unbatched 51ms
        If 2000 extra key-values are migrated: unbatched 66ms
        If 5000 extra key-values are migrated: unbatched 113ms
        If 8000 extra key-values are migrated: unbatched 162ms
Assuming 5000 key/values touched by block execution:
        If 1000 extra key-values are migrated: unbatched 93ms
        If 2000 extra key-values are migrated: unbatched 110ms
        If 5000 extra key-values are migrated: unbatched 157ms
        If 8000 extra key-values are migrated: unbatched 207ms
```

## Benchmark-style measurement

In the previous PR, I also introduced a more formal benchmark.

Before (using new-ish batched API) took ~101ms per op:

```bash
$ go test . -run=none -bench=BenchmarkBatchLeavesInsert -memprofile=mem.out -cpuprofile=cpu.out
goos: linux
goarch: amd64
pkg: github.com/gballet/go-verkle
cpu: AMD Ryzen 7 3800XT 8-Core Processor            
BenchmarkBatchLeavesInsert-16                 10         100468352 ns/op        51978069 B/op     124714 allocs/op
PASS
ok      github.com/gballet/go-verkle    3.056s
```

After (this PR) the use of plain `Insert(key, value)` takes ~102ms per op:

```bash
$ go test . -run=none -bench=BenchmarkNewLeavesInsert -memprofile=mem.out -cpuprofile=cpu.out
goos: linux
goarch: amd64
pkg: github.com/gballet/go-verkle
cpu: AMD Ryzen 7 3800XT 8-Core Processor            
BenchmarkNewLeavesInsert-16           12          98242143 ns/op        49037674 B/op      87698 allocs/op
PASS
ok      github.com/gballet/go-verkle    3.639s
```

**TL;DR:**

- **Synthetic benchmarks: as fast as the specialized API.**
- **Benchmark: as fast as the specialized API, and some fewer allocs.**
- **API surface reduced again to usual APIs.**
- **Not only can overlay-tree key/val migration use the usual APIs, but block-execution cases of `Insert(..., ...)` that create new LeafNodes will have a faster tree `Commit(...)` performance.**

I tested this branch in the replay benchmark and got a similar performance. I inspected why, and it’s because all block executions only have a few new leaves created, so this optimization of batching newly created leaves isn’t exploited that much. Probably when importing “heavier” blocks and being post-byzantium, we could appreciate the difference more. It might still be useful to double-check our current replay benchmark in the reference machine.